requirements.txt
flask>=2.0.0
numpy>=1.21.0
opencv-python>=4.5.0
opencv-contrib-python>=4.5.0
Pillow>=8.0.0
pytesseract>=0.3.8
scikit-learn>=0.24.0
scipy>=1.7.0
scikit-image>=0.18.0
matplotlib>=3.4.0
werkzeug>=2.0.0

